requirements.txt
tqdm
ipywidgets

matplotlib
seaborn
plotly

numpy
pandas
scikit-learn
scipy
opencv-python

jupyterlab

jupyter-ai
jupyter_ai_magics
langchain-openai

simple-pid

pygame
pyserial 

pynput
